Fixed #3
Added a chronometer to the teleoperated and autonomous score fragments.
Added a warning for more discs collected than obtained.
Added a warning for more than four discs held.
Added a warning for collecting an unreasonable amount of discs.
Added a warning for being on a screen after the match has ended.